3/16/2024 changes by Oxun.

+++ src/main/java/uz/tsx/repository/AnnouncementRepository.java
@@ -5,7 +5,7 @@ import uz.tsx.entity.announcement.AnnouncementEntity;
 
 import java.util.List;
 
-public interface AnnouncementRepository extends JpaRepository<AnnouncementEntity, Integer> {
+public interface AnnouncementRepository extends JpaRepository<AnnouncementEntity, Long> {
     List<AnnouncementEntity> findAllBy();
 
 }
